Fix issue with Date.today and Date.clone
That were somehow missed in a previous spectest  :-(

Date.today called without a :formatter argument, would initialize
the &!formatter attribute without a container, which would cause
a very obscure issue in the dispatcher when trying to .clone such
a Date object *with* a :formatter argument specified.

Date.clone allows twiddles, but didn't check any :day argument
for non-Intness.  It does now.

Part of work on #1238