Lose Positional argument to BUILDALL
This makes Date instantiation (other than Date.today) about 45% faster,
because it uses the direct POPULATE interface.  And drops the spectest
time from 211 -> 208 wallclock (1308 CPU -> 1279 CPU seconds).

The standard object instantion uses the following call sequence:

.new -> .bless -> BUILDALL

However, for historic reasons, the call to BUILDALL also passes a
Positional as the first argument, but does not use it.  This artefact
is from an era when it was considered to have some automatic handling
of positional arguments to .new.  This idea has long been abandoned
since then, but the Positional interface was still part of every
object instantiation.

The easiest solution would have been to just remove that Positional
argument everywhere in core.  However, quite a few modules actually
depend on this historic interface (Inline::Perl5 being one of them).

So I decided to rename the method internally to POPULATE, which just
takes a hash with named arguments.  And keep a BUILDALL method around
with the old interface, passing its hash on to POPULATE.  So the
call sequence is now:

.new -> .bless -> POPULATE

Since BUILDALL is tested for in some (spec)tests, some changes are
going to be needed there as well (once 2025.01 has landed).