Make reduction of hyperop work on 0 / 1 arguments
This was first noted in

  https://github.com/Raku/old-issue-tracker/issues/3475

in 2014.  Its partial fix in the end caused inconsistency in handling
reduction of operators with left associativity, which was reverted
in https://github.com/rakudo/rakudo/commit/9bcb01e966.  This now
fixes the original issue, at least until we have a better way of
handling hyperops (which are currently codegenned to dispatch to
a HYPER sub).

This fixes two TODOed tests.