Throw coercion failures on object hashes
If an object hash is defined with a coercion type (e.g. Int(Str))
the Failure of the coercion would be stored as the key.  Only to
thrown whenever that key would be accessed in the hash, e.g. with
.keys.

This commit checks for Failure inside the storing methods of object
hashes, and throws that immediately instead of just storing the
.WHICH of it.

Inspired by #4301

Spectests clean.