RakuAST: cope with exceptions from ACCEPTS in check for duplicate multi sigs
Trying to match signatures at compile time can lead to exceptions, i.e.
when one or both of the signatures contain types with customized ACCEPTS
methods. In this case we just cannot know whether the signatures are
identical, thus we should fail open and assume that the programmer knows
what they're doing.