Fix issue with .item causing PDF failure
I don't like this solution at all.  Historically, the concepts of an
item and a container have been conflated, which they probably shouldn't
have been.

So now [].time will return an unassignable container, because there
are several tests that depend on that:

  - t/spec/S06-traits/misc.t
  - t/spec/S17-lowlevel/cas-loop.t
  - t/spec/S17-lowlevel/cas.t

which feels inconsistent to me.  Also, I don't see how the code in
Iterable actually returns a unassignable container: there must be
some magic going on behind the scenes at the QAST or possibly MAST
level.

We should probably re-visit the whole concept of what a container
is, and what an item is for 6.e, or later..