RakuAST: preserve proper nesting of blocks
Previously we added all blocks as declarations of the surrounding
lexical scope. This worked well enough because most times it's a lexical
scope that generates a block. However this is not always the case.
Thunks also generate blocks. Problems start when we e.g. closure clone
such a thunk. If blocks belonging to its children are not properly
nested, the closure clone won't cover them. Then we end up with invalid
values in variables.

Fix by differentiating between lexical scopes and block creators.