Make sure result of >>foo<< is stored correctly
By default, hyper will create the same data structure as it finds
on the left side.  However if the result can not be stored, it would
die with a typecheck error.

Now it will return a List with the result instead if the storing
of the values in the same data structure failed.

Partial fix of #5778