Make sure that .deepmap returns the right structure
And make sure that it is filled in the correct way.  Previously
it would just hack the IterationBuffer into an object of the same
type as the invocant, which allowed wrong values to appear inside
the data structure, as was spotted in #5778.

This also reverts 1aea298ded and fixes it in the same manner: it
should first be checked if it can call a "STORE" method, and *then*
doing setup: originally it would just return 0 if there was no
"STORE" method on the object.

Whether checking for a "STORE" method is a correct heuristic, is
debatable: for now, it appears to cover all possible (native) array,
hash and QuantHash cases.