Introduce Hash::Ordered
This subclass of Hash will keep the order in which keys have been
added, with as little overhead as possible.

The theory of operation is that whenever a key is accessed, and the
key does **not** exist in the hash, then the key will be added to
the list of known keys.  If a key is deleted, that key is **not**
removed from the list of known keys (but it *is* deleted from the
hash).

Only when some form of iteration is needed over the keys (or values)
of the hash, is the list of known keys (which is in order of appearance)
checked for existence in the hash, and any double entries or deleted
keys are removed, and the list of keys updated.

The current syntax is:

  my %h is Hash::Ordered = ...;

This could probably be improved by either a dedicated trait, such as:

  my %h is ordered = ...;

or maybe some other syntax, such as:

  my %h{:ordered} = ...;

Support for typed ordered hashes is basically automatically available.

Support for object hashes would need to be made separately.  This
shouldn't be too hard, but felt only necessary if this change to the
Raku Programming Language is approved.

Also, maybe this should become a 6.e feature: but then integration into
the legacy, as well as the Raku grammar, would be more troublesome.