Handle single param to named array args better in MAIN
Before this, the handling of named array args was to first just
try to bind the given args to the MAIN signatures.  If that failed
it would convert all named parameters to lists in the capture,
and then try again.

This had a side-effect as described in #5159: if this convesion
happened, all *other* named arguments would be represented by
arrays as well.

Instead of just converting all arguments to Arrays in the capture,
it tries to convert all **combinations** of named arguments to
Arrays, and the first combination that works, will be accepted.

Since the number of named arguments is not likely to become very
large, it feels like this won't be suffering from a combinatorial
explosion any time soon.