Fix multi-dim array slices with lazy indices in 6.e
Before this commit, something like @foo[1..*; 2] would hang because
it would continue to produce values for index 2 on Any as soon as
the first dimension was exhausted.  This was reported in #5122.

This fix consists of checking whether an outer index is a lazy
Iterable.  If so, it will install a $*LAZY dynamic variable, which
will be checked when producing values one level deeper.  If an
empty slot is encountered, and $*LAZY is set, then it will be
considered the end of iteration at that level.

This also fixes all the cases with adverbs as well.