Make Blob ~ Blobx take lefthand type in 6.e
Reported in #5071: having it produce a Buf if the types on either
side are differently parameterized, felt weird as in other such
cases the type on the left-hand side of the infix will be produced.

This is now the case in 6.e.  Sadly, the revision-gated mechanism
could not be used.  Adding it to the proto broke setting compilation,
and applying it in the core epilogue made the core setting compiliation
infiniloop (probably because it is using infix ~ somewhere inside).