Fix a bunch of non-sunk lazy loop bugs
Whether a loop used as expression turns out to become lazy or not has so
far been determined by the complex interplay between several bugs. We
treated loop labels as conditions, did not recognize 1 as a compile time
value and failed to make iterators lazy. This fixes all of those.