RakuAST: forbid interpolating attributes in regexes
We forbade interpolating attributes in regexes, because regexes are
really methods and have their own self, thus the attribute would refer
to a (likely non-existing one) on the Cursor. Since there seemed to be
no way to make this do what the user probably meant, it was deemed
better to forbid this usage altogether.

Ironically the RakuAST frontend would actually do the impossible. But
there are spec tests requiring us to forbid this anyway. Well, it's easy
to remove these checks if we want to re-visit that decision.

Fixes: /$!a/ # to fail with an appropriate exception