Make sure grammar changes persist in the REPL
Fixes #2245.  This fix consists of 3 elements:
1. Trying to set a $*GRAMMAR dynamic variable if a categorical is
   added to the grammar.  This will generally fail, because no
   $*GRAMMAR variable is visible by default.
2. Ensure that the current grammar and actions can be determined
   with nqp::gethllsym (assume Perl6::Grammar/Actions for now)
3. Add $!grammar and $!actions attributes to the internal REPL class
   to be set at instantiation, and specify these in the .eval call.
   Create a $*GRAMMAR dynamic variable in scope of the compiler and
   check it after the eval whether it is now set.  If set, adapt the
   the $!grammar and $!actions attributes accordingly

This adds minimal overhead at startup (2 extra nqp::bindhllsyms),
only minimal overhead when a categorical is added (a try bind), and
some overhead for each expression in the repl (checking if the
grammar has changed).

All in all surprisingly little code was needed for an issue that is
now about 8 years old.