Simplify solution for #2245
It occurred to me that we don't need to know the grammar /actions
normally in the REPL, only if a change was made by adding a
categorical.  So it's just needed that the grammar /actions
arguments are *not* passed to the compiler if they haven't
changed yet.

This also has the advantage of not needing any changes once
the Raku grammar becomes the default.