RakuAST: fix scope confusion when adding categoricals
Only when add-category is called while we're in the new operator's block
should we add %?LANG to the outer scope. In all other cases it needs to
be installed into the current scope. Otherwise we risk confusion when
categoricals are added in multiple blocks and then someone does an EVAL.