RakuAST: fix ::Literal.from-value for general values
Before, if the WHAT of the value did not point to a known class
(such as a Bool) then it would instantuate a Mu object and try
to set the attributes in that.  Now check for the existence of
the class first, and if it doesn't exist (and so doesn't have
any special handling) make a generic RakuAST::Literal object
(which QASTs to a WVal).

Found while working on #2599, making .assuming work using RakuAST