RakuAST: fix bogus parse error when trying to call trait_mod:<is>
When trying to determine whether a name is a meta op, we call back into
the parser to just try and see whether we find any match. However when
the parser throws a panic, this does not mean that we have to abort. It
simply means that it's not a meta op at all.