Re-imagine .assuming using RakuAST
This completely removes the old implementation of .assuming by
a RakuAST based implementation.  For the users, there should be
no noticeable difference, apart from:

- slightly faster creation of new Callable
- 2.5x to 50x faster execution of the created Callable

Instead of creating a subroutine that does all sorts of things
during runtime, the RakuAST approach allows for the creation of
a very thin shim, which will possibly allow inlining if the original
sub wasn't too large.

For example:

    sub foo(Int $a, Int $b, Int $c) { dd :$a:$b:$c }
    my &bar = &foo.assuming(42);

then the resulting subroutine would effectively be created as:

    sub bar(Int $b, Int $c) { foo(42,$b,$c) }

Inspired by #2599