Remove leap second check step from release process
It looks like there won't be another leap second *added* in our
lifetime.  There might be a negatve leap second, but that would
require quite a bit more than just running the leap day script.

In any case, having another leap second added, will be major news
in the scientific world.  So most likely it will filter down to
the Raku core devs, and one of them can then run the script with
the correct date.

In light of #5769