RakuAST: fix type mismatch in supply setup code
The sigil on &*WHENEVER-BLOCK-TO-ADD says that its value must be a
Callable. For unknown reason the old compiler frontend didn't generate a
type check for the bind, so it was ok with having Nil bound to the
variable. However that's highly confusing and plain wrong as its
breaking the contract. Actually the code using this variable expects it
to be a Block, so use the Block type object as the initial value.