RakuAST: replace Perl6::Metamodel with just Metamodel in the setting
After the BOOTSTRAP the Metamodel namespace should be renamed. Outside
the setting these classes have always only been accessible through
Metamodel (not Perl6::Metamodel). It seems to be an implementation
accident that they were available as Perl6::Metamodel in the setting.
There is no advantage to using that name and its not available when
compiled using the RakuAST frontend. Most likely because that doesn't
load Perl6::Grammar or the like, i.e. classes in that same namespace.