RakuAST: fix while/loop loop statement mods when we skip CHECK time
StatementModifier::WhileUntil needs us to thunk loop-condition, condition-
modifier and expression but crucially it only needs this if we're not a block
statement and we're not sunk. The problem is that at BEGIN time we don't know
that yet.

Previously we therefore thunked at CHECK time, but we may not even get a
CHECK time if the statement is executed at BEGIN time. Thus we now thunk at
BEGIN time just in case and if it turns out the statement is sunk, we undo
that thunking again.