RakuAST: add lexical fixups for role bodies
When code runs at BEGIN time, such as role bodies and BEGIN blocks, we need
to ensure we get lexical outers fixed up properly when deserializing after
pre-comp. To do this we make a list of closures, which each point to the outer
context. These survive serialization and thus point at what has to be fixed up.