RakuAST: allow literal expression such as 2³⁺²ⁱ
In response to #4465.  Felt that the most sense it would make to
allow val() based expression in so far the superscript characters
allow that (since there is no superscript period or e, this disallows
Rats and Nums).

- adapt the <power> token to allow val() based expression
- add helper method in Fixups for convenience to access val() logic
- adapt associated action to use that helper method
- fix deparsing

Since this feature is *not* used in the core C setting, it is safe
to put the helper method in Fixups.