Wrap enum value stringification in a quietly
In response to #5794.  Enums keep both a forward hash as well as
an inverted hash.  The issue describes warning creating the reverse
hash if type values were encountered in data-structures such as
hashes and arrays used as value in an enum.

Since this needed a fix in the metamodel, and NQP doesn't know
quietly, basically copied the QAST from the quietly handling in
the grammar.