RakuAST: don't add role stubs to their role groups
Normally when a package is stubbed, the generated meta-object is later
re-used when the package is fully defined. This ensures that any lookups
done between stubbing and defining ultimately resolve to the meta object
that is fully defined.

With roles however we do not actually install the role's meta object in
the lexical and global scopes. Instead we install a role group. We later
add the specific role to as a possibility to the role group. Thus when
we check whether a stub already exists, we would not find the stub, but
this role group. We obviously cannot use the group as meta-object.

Luckily we don't care too much about that stub's meta object and can
simply ignore it. We don't add the stub to the group's possibilities and
are then free to create a fresh new meta object when we actually define
that role.