RakuAST: fix while/loop statements when we skip CHECK time
Similar to commit 4a8ee5d22d9745153780bd895dd7125cf034872b

While/until loops need to thunk their expression/inrement,  but crucially only
if they're not a block statement and they-re not sunk. The problem is that at
BEGIN time we don't know that yet.

Previously we therefore thunked at CHECK time, but we may not even get a
CHECK time if the statement is executed at BEGIN time. Thus we now thunk at
BEGIN time just in case and if it turns out the statement is sunk, we undo
that thunking again.