Preliminary fix for .assuming producing typed slurpies
Problem became apparent after 518596b47ee659a8 but did not appear
to create any problems in roast before that.

There is still a kind of tangled mess in RakuAST land handling
typed @ and % sigilled parameters, because internally this is
handled as Positional[type] / Associative[type], but in other
places the sigil appears to be leading, causing stacked types
such as Positional[Positional[type]] in certain situations.