RakuAST: don't sink expressions with loop modifiers
For unknown reason we must not sink the bodies of loops created from
statement modifiers. This restriction may not even have been
intentional, but it's certainly being relied upon by code in the setting
like this from List.Capture:

nqp::istype(($v := nqp::atpos($!reified, $i)),Pair)
    ?? nqp::bindkey($hash, $v.key.Str, $v.value)
    !! nqp::push($list,$v)
    while nqp::islt_i(++$i,$elems);

By any logic this should sink the value return from bindkey and push
(which would be $v), but we don't and if we did and $v is a Seq this
would consume the iterator we're trying to bind to a signature.