RakuAST: fix access to QRegex and QAST::Regex in the setting
The old compiler frontend loads QAST and QAST::Regex in the Actions
which somehow makes them available to the setting code itself. The
RakuAST frontend doesn't, so we need to load them explicitly. However in
"early" setting code, module loading is not yet operational. We
therefore load them in bootstrap and make them available as hll symbols.