RakuAST: ensure sink is calculcated when generating QAST for subs/methods
We really need to have sink calculated before we generate QAST as it can
gravely affect compilation output, by e.g. turning lazy loops into eager
ones. In normal compilation flow this isn't an issue, but of course
routines can be called at BEGIN time which means we also generate the
QAST before CHECK time (at which we normally would calculate sink).

So force calculation when we generate QAST. We already did so before but
just for role bodies (as those are always compiled at BEGIN time). Now
we extend that to all subs and methods.