RakuAST: only generate return handlers when there's a chance of explicit return
Routines only need a return handler if it's actually possible that an
explicit return is performed. Unfortunately finding out with certainty
whether that's the case is impossible because there are many ways of
calling return including method calls and from EVALed strings. Thus we
need to be prudent and add the handler if there's even a possibility
that it can happen.

Example of such a worst case:
sub foo() {
    my $ret = 'ret'; my $urn = 'urn';
    "return-value"."$ret$urn"();
}