Expand Hash parameterization
This allows one to specify a default value for any containers as
well.  By default, this is the same as the "of" type, as it was
before.

So the standard hash "parameterization" is now Hash[Mu,Str(Any),Any]
(as it was before, but now you can actually explicitely parameterize
a hash that way).

This commit does not actually change anything in the way object
hashes are created.  That will be next.

This breaks two .assuming tests in t/spec/S06-currying/positional.t
that will need fixing.