RakuAST: fix detecting number of whenevers in a supply
The previously used logic to detect whether a supply has a single
whenever was inadequate to the complexities of Raku code. It only
detected the straight forward cases of multiple whenever statements
directly in the supply block or one whenever directly in the
block of another whenever. There are however constructs like:
$current = do whenever inner
which are harder to detect.

Luckily we have the Attaching mechanism for reliably finding arbitrarliy
deeply nested nodes.