RakuAST: ensure we decont the result of parameter coercion
During parameter processing we set up a local variable for the parameter
and also a deconted one. In some cases like with coercive parameter
types we use that deconted value and produce yet another value. In the
compiler code we then clear $get-decont-var to signify that it no longer
refers to the most current version of the parameter value (that would be
the coercion result which we write to the original local variable).
However we do need to decont the result of the coercion as well but
neglected to so so.

Fix by issuing a decont op, if $get-decont-var does not give us a decont
var.

Fixes: sub foo(int $tz) { }; sub bar(Int(Cool) $tz) { foo($tz) }; bar($*TZ)