Make sure Hash/::Typed/::Object take same parameterization
Not entirely sure exactly what is going on here, but .assuming got
confused by the fact that Hash::Typed only took 2 paremeterization
arguments, and Hash and Hash::Object take 3.

Code should just use the Hash[...] parameterization.  But internally
this parameterization is used to parameterize one of two roles:
if the parameterization for key is Str(Any), then Hash::Typed will
be parameterized and mixed into Hash.  Otherwise, Hash::Object will
be parameterized and mixed in.

What was happening, is that .assuming tried to parameterize
Hash::Type with the arguments it got for Hash[...], which did the
wrong thing.  By making sure the positional parameters are in the
same order and have the same meaning, we're working around this
issue.

I guess in the future, there's something to be said for moving to
named arguments for "of", "keyof" and "default" in parameterization.

This fixes the two failures in t/spec/S06-currying/positional.t
introduced by bc05703131b6671608