Fix Buf/Blob.join/raku for uints
Currently there appears to be a deficiency in the automatic coercion
of a native 64bit integer with the most significant bit set, to a
native string.  This is probably because there is no special handling
for that, and so the signed integer handling is used.  Which isn't
a problem if the msb is not set.

Workaround this deficiency by using a HLL .Str for unsigned integer
Buf/Blobs in .join, which forms the base of .raku.

Spotted by Lucien Grondin++ at
  https://irclogs.raku.org/raku/2025-03-20.html#11:15-0003