Restore the old implementation of .assuming
A blin run made it clear that quite a few modules in the ecosystem
depend on .assuming being used in a pre-compiled module environment.

Alas, this appears to be incompatible with the RakuAST based .assuming
while running with the legacy system: it looks like it is going to
need the RakuAST bootstrapped rakudo to be able to do that reliably.

So until then, remove the old, restore the new  :-(

Reverting this commit should be enough to restore the RakUAST based
version of .assuming again