RakuAST: fix calling post-declared subs from role methods
When a Role's method tries to call a sub that is declared after the role
itself, the lookup for this routine would fail. That's because role
bodies are compiled immediately after the role is parsed. At this point
the callee does not exist yet, so we won't find it. However at runtime
the lookup may very vell succeed.

So to get this to work we need to:
* stop running CHECK as not all information is available to do it
  properly
* just generate a lexical lookup for the originally requested name if we
  cannot resolve a named call
* fix accidentally putting a Nil in place of the call