RakuAST: fix mis-parse of colonpairs in type parameters
Apparently the author of the code thought that $*IN-TYPENAME would be
scoped only to the branch it was declared in, but it actually covered
the whole token. This means that the variable was also set when parsing
type parameters which made us produce a different result for colonpairs.
This in turn prevented us from determining their compile time value
making a `does Foo[:a(2)]` fail at compile time.