RakuAST: fix role parameters not found in where clause in role signature
We only declared lexicals for role parameters after entering the role
body. That's too late when those parameters are referenced from within
the role's signature. We previously didn't declare those lexicals on
parse because the resolver was still in the role's outer scope. As a
workaround we declared those lexicals manually after entering the role's
scope.

Fix by entering the role's scope earlier, declaring the lexicals
normally as part of parsing the signature, then leaving the scope again
and re-entering in the begin handler. Could be that we can leave out the
dance, but this at least lets us progress now and as workarounds go is
not really uglier than what we had before.

Fixes: role Foo[$x where { $x }] { }