RakuAST: tighten up handling of declarator docs
Declarator docs need to be on consecutive lines: no (empty) lines
inbetween are allowed, or the declarator doc will be considered to
be without a declarand and cause a worry.

The first whitespace character after a #= or #| is ignored, but does
not have to exist.  So this allows '#=' as content of a whole line,
but also '#=foo' to be generating the "foo" string.