RakuAST: re-structure package installation sequence
Packages must be installed (i.e. their names declared) before the parser
enters the package's scope. However we can only install a package after
parsing the traits as for the declaration we need the meta object stub
and for that we need to know the repr which can be changed via trait.
Thus parse time would be too early.

Split off package installation into a separate method that will be
called by the parser and also by PERFORM-BEGIN (but only have an effect
once). This gives us the extra step we need to avoid having special code
for parsing in the BEGIN handler.