RakuAST: report trying to call unknown routines in BEGIN blocks
If we encounter a call to an unknown routine in BEGIN time code, there's
two cases:
* the call is in some nested code block
* the call is at the top level of the BEGIN block

If it's in a nested code block, it's possible that this code block will
not be executed immediately and that the missing routine will be
declared between execution of the BEGIN block and execution of that
nested block. Thus we keep runtime resolution of that name in place and
hope for the best.

If however the call is at the top level, we know that it's going to be
executed during this BEGIN phaser. Thus if we haven't found the routine
so far, it's not gonna happen in time. So report that with a proper
exception instead of getting a much less helpful "cannot invoke object
of type 'VMNull'".