RakuAST: don't sink operands of operators (except last one on short-circuit)
For a simple `1 + 2` we don't want warnings about 1 and 2 individually,
but instead on the whole expression. After all, those numbers are used
to calculate the sum and it's that sum that's then unused.

Could be however that we're gonna need to pull up the needs-sink-call
result from operands to the infix, i.e. call sink on the infix' result
if any of the operands requires a sink call.