RakuAST: only set declarand for attributes
Setting declarator docs on variables in general is currently impossible
without a significant overhaul of the way declarator docs are being
handled: currently, object types that support declarator docs have a
$!why attribute, which is usually not used.  An approach with mixing
in a $!why attribute on demand on Scalar containers proved to be a
failure.

So all variable declarations (apart from attributes) are now ignored.