RakuAST: add bare dd CHECK functionality
As a core developer, I'm interested in the exact RakuAST tree that
is being built for a given piece of code.  Until now, I usually
resorted to:

    raku -e 'say Q|the code|.AST'

which is pretty cumbersome if you have code with | in it (because
then you need to use another delimiter for Q).

Then I realized that the CHECK phaser exposes $*CU, which contains
the complete RakuAST tree.  So I figured that "dd" being called inside
a CHECK phaser *without* any arguments should DTRT.  So now I can say:

    RAKUDO_RAKUAST=1 raku -e 'the code; CHECK dd'

to get the same result: showing the RakuAST tree.  But that's more
typing, you say.  Indeed, but I also have a "ru" alias:

    alias ru='RAKUDO_RAKUAST=1 raku -e'

which allows me to do:

    ru 'the code; CHECK dd'

to make life easier for me, and hopefully for other people.