RakuAST: fix unwanted redeclaration error on my scoped multi-part-name packages
A my-scoped package with a multi-part name like X::Await::Died pus into a bit
of a conundrum. On the one hand it should only be visible in the lexical
scope it was declared in, as the my-scope requests. On the other hand
packages for a namespace hierarchy and the "Died" class we declare
should be available in X::Await's stash. Thus if the root package
already exists in some outer scope, as is the case for X in the CORE.d
setting, we silently change the package's scope to our. This however
made us throw a Redeclaration exception, as X::Await::Died does already
exist in that stash, since it's declared in the CORE.c setting.

To get the CORE.d setting to compile again, we no longer consider this
case a redeclaration, i.e. if my scope was requested, the existing
package gets overwritten silently.

The real fix would be to either prohibit my scoped declarations with
multiple names, or to create a lexically scoped proxy version of the
parent package that knows about the new package. Both are major
undertakings though and have nothing to do with RakuAST itself.

This also unbreaks t/12-rakuast/types.rakutest