RakuAST: make warnings test more lenient
In the old frontend the optimizer collects warnings about unused values,
groups them by message and appends the list of line numbers for each.
All of this is manual code just dealing with message strings. The
RakuAST frontend however builds an exception for each occurence and adds
them to an X::Comp::Group which is arguably the more structured
approach. However this path does not do any deduplication. In both cases
the same information is conveyed. It's just a more compact format vs. a
report that's strictly in order of source location. So instead of trying
to re-create the exact behavior of the old implementation, make the test
a little more lenient. It is after all just about whether the
information is there, not what the exact format is.