RakuAST: defuse overly specific test
The RakuAST frontend throws a more tightly typed exception (with exactly
the same message). As the important part is the message and to keep the
old frontend passing the test also, reduce the constraint to just any
exception. Expecting an AdHoc specifically doesn't make sense anyway as
it's preventing us precicely from introducing a more targeted type.

lizmat++ for the suggestion