RakuAST: better fix for clobbered $_ in EVAL
With commit ede438efa8e83b66e4286ab1bdff1c236e7e95c3 we started
declaring a $_ in an EVAL to stop the outer $_ from getting clobbered
e.g. by an exception in a smartmatch and the resulting failure to reset
$_ to the previous value. However this meant that an outer $_ was no
longer visible in an EVAL which broke other tests.

The proper fix is to declare that $_ but to also initialize it with the
outer $_'s value. If it then gets rebound e.g. for a smart match, only
the local variable will get rebound and the outer $_ retains its
container.