IO::Handle.slurp: Read entire file contents at once
We can easily pre-size the read volume in the call to nqp::readfh
for the cases where it makes sense (ie., every path that isn't
"-".IO).

We still preserve the older behavior in case the byte length is
super low or if there is some other interruption to getting the
self.s value at runtime.