Remove type specification by name in Parameter constructor
The Parameter constructor has a "type" argument that besides type
objects also accepts concrete values. In most cases it uses that
value's type but if the value is a string, it instead checks whether
that string might be a type name and looks up that type instead.

By special casing strings ending in ')' this feature violates the
principle of least surprise. If the string ends in ')' but does not
contain a type name, it throws an internal error.

This feature is untested, undocumented, broken and unwise, so let's remove it.
Nowhere else do we allow for specifying types by their name and for good
reason. Type names are not unique and they can become quite complex
containing even arbitrary code in the case of parameterized roles.