Small revision to the fix for R#5875
Unlike the PseudoStash lookups, we do not need to use `try` to
guard the lookup into the `%caller-my` hash.

Also, the previous phrasing was actually equivalent to:

    try (%caller-my{a} // ::(a))

Which was probably fine in practice but it was less clear than
the revised version where we leverage the fact that asking a
hash for a missing key will never produce an exception.