Enhance feedback around named subset parameters
Adding a worry here provides users with a bit of
helpful feedback when they define an optional
named parameter with a subset type but fail to
provide a valid default value.

Default values are also checked for conformance
to the provided subset, with the obvious caveat
that an expression without a compile time value
will pass through this protection.