Return to original approach of triangle METAOP_REDUCE_LIST
As described in #5882, commit ba2431f4456ad43 actually changed the
internal logic of the triangle metaop list reducer to check for
the first value: if that indicated the end of the values, then
the iterator wouldn't actually be returned, but an optimized version
for empty lists.

This caused an issue with infix:<,>, because infix:<,> returns a
List even for a single element, which would then become the first value.

This commit reverts that approach by *not* specifically checking
for empty lists as input, as was the case before (although that used
gather), but instead check for that on each pull-one instead.

This significantly simplifies the iterator: don't know what I was
thinking setting it up the way it was, as the (probably very meager)
optimization for empty lists probably wasn't very noticeable indeed.

Fixes #5822.