Change Linux pre-built release pipeline, CentOS 7 -> Ubuntu 16.04
Azure removed support for the Ubuntu 20.04 agent OS. We need to
mount the docker binary into the container to work around missing root
permissions in the container. That only works when the host's docker binary
is glibc compatible with the container glibc. Upgrading Ubuntu from 20.04
to 22.04 made the host's docker binary incompatible with CentOS' 7 glibc.
The same problem arose with several other distributions and glibc versions.

So as a fix, I created a derived Ubuntu 16.04 container that already has
`sudo` preinstalled. This alleviates us of the need to run the host's
docker in the container and resolves the above issue. The upgrade of the
container OS to Ubuntu 16.04 was not strictly necessary. But given I was
already fuddling with the pre-built release pipeline, this felt like the
right time to do so.
An important effect of the OS upgrade is that the minimum required glibc
version for Rakudo is now changed from 2.17 to 2.23.