Give $*KERNEL.raku a more meaningful representation
Before this commit, $*KERNEL.raku would show:

  Kernel.new(release => Str, hardware => Str, arch => Str, bits => Int, name => "unknown", auth => "unknown", version => Version, signature => Blob, desc => Str)

which isn't very informative.  The reason for this behaviour is
that the setting of the attributes is as lazy as possible, and
the standard .raku provided by the systemic role just looked at
the attributes.

This commit introduces a local Kernel.raku method that instead will
use the .raku presentation of each of the accessor methods of the
attributes to set up the representation.  Which for me now shows:

  Kernel.new(name => "darwin", version => Version.new('Darwin.Kernel.Version.24.5.0.Tue.Apr.22.19.53.27.PDT.2025.root.xnu.11417.121.6.2.RELEASE_ARM.64._T.6041'), hardware => "arm64", arch => "arm", bits => 64, release => "24.5.0", desc => Str, signature => Blob)

which is *way* more informative.