RakuAST: introduce $?SOURCE and $?CHECKSUM
This is achieved by adding a '$!source` attribute to ::CompUnit
and add the $?SOURCE and $?CHECKSUM variables in the add produce
implicit declarations logic.

$?SOURCE contains the source code that produced the bytecode that
is executing this: inclusion can be prevented by specifying the
RAKUDO_OMIT_SOURCE environment variable with a true value.

$?CHECKSUM contains a SHA1 of the source code of $?SOURCE.  This
is always available.

Intended to make it easier for the Rakudo debugger to do its work.