Fix weird interaction in 6.e with string concatenation in ^parameterize
Methods with a ^ at the start of the name are weird in the sense
that they don't exist as a method of the class itself, but that
they're mixed in the metamodel class of the class (which is usually
written in NQP).

In 6.e this somehow causes a problem when a HLL literal string, such
as '[' or ']' is passed as a parameter to infix:<~> when it is
called.  This shows up as:

  ===SORRY!===
  lang-call cannot invoke object of type 'VMNull' belonging to no language

with the line number pointing to the line doing the concatenation,
and down from there, the call to the '^parameterize' method.  Not
very informative.  Even with --target=ast  :-(

This commit works around this issue by using nqp::ops to concatenate
the literal strings in the ^parameterize methods of the QuantHash
classes before joing them together with infix:<~>.