RakuAST: add RakuAST::Enum class for handling enums
Intended to be used for core enums, but could also be used for
custom ones.  Is in fact a very shallow subclass of ::Term::Name,
but allows for better deparsing and better readability.

Also added a "from-identifier" instantiator method to ::Term::Name
to make it easier to create ::Term::Name objects that have simple
identifiers, such as "True" and "False"